2. Basic annotations - floor

+++ 0x00-python_variable_annotations/2-floor.py
@@ -11,4 +11,4 @@ def floor(n: float) -> int:
     Returns:
         int: the rounded number
     """
-    return floor(n)
+    return int(floor(n))
